requirements.txt
PyYAML
streamlit
streamlit-ace
dask[complete]>=0.13.0
graphviz>=0.4.9
matplotlib>=1.0.0
numpy>=1.6.0
pandas>=0.15.0
python-dateutil>=2.0.0
scipy>=1.4.1
sympy>=1.0.0
numpydoc>=0.6.0
sphinx>=1.4.1
sphinx_rtd_theme>=0.2.0
cython>=0.29
pygom
google-cloud-bigquery>=3.5.0